rename the extension
