+++ src/mdp/azext_mdp/aaz/latest/mdp/__cmd_group.py
@@ -12,7 +12,7 @@ from azure.cli.core.aaz import *
 
 
 @register_command_group(
-    "dev-ops-infrastructure",
+    "mdp",
     is_preview=True,
 )
 class __CMDGroup(AAZCommandGroup):
