+++ src/mdp/azext_mdp/aaz/latest/mdp/pool/_delete.py
@@ -12,7 +12,7 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "dev-ops-infrastructure pool delete",
+    "mdp pool delete",
     is_preview=True,
     confirmation="Are you sure you want to perform this operation?",
 )
