+++ src/mdp/azext_mdp/aaz/latest/mdp/pool/_update.py
@@ -12,7 +12,7 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "dev-ops-infrastructure pool update",
+    "mdp pool update",
     is_preview=True,
 )
 class Update(AAZCommand):
