+++ src/mdp/azext_mdp/aaz/latest/mdp/pool/_wait.py
@@ -12,7 +12,7 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "dev-ops-infrastructure pool wait",
+    "mdp pool wait",
 )
 class Wait(AAZWaitCommand):
     """Place the CLI in a waiting state until a condition is met.
