+++ src/mdp/setup.py
@@ -34,16 +34,16 @@ with open('HISTORY.rst', 'r', encoding='utf-8') as f:
     HISTORY = f.read()
 
 setup(
-    name='devopsinfrastructure',
+    name='mdp',
     version=VERSION,
-    description='Microsoft Azure Command-Line Tools Devopsinfrastructure Extension.',
+    description='Microsoft Azure Command-Line Tools Mdp Extension.',
     long_description=README + '\n\n' + HISTORY,
     license='MIT',
     author='Microsoft Corporation',
     author_email='azpycli@microsoft.com',
-    url='https://github.com/Azure/azure-cli-extensions/tree/main/src/devopsinfrastructure',
+    url='https://github.com/Azure/azure-cli-extensions/tree/main/src/mdp',
     classifiers=CLASSIFIERS,
     packages=find_packages(exclude=["tests"]),
-    package_data={'azext_devopsinfrastructure': ['azext_metadata.json']},
+    package_data={'azext_mdp': ['azext_metadata.json']},
     install_requires=DEPENDENCIES
 )
